backend/requirements_facial.txt
google-generativeai>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
pillow>=10.0.0
python-dotenv>=1.0.0
